device: drop NMDeviceVlan:bring_up() implementation
This retry loop was added by commit dc6341acec2d12856265d6976420d56f0560410b.
But I suspect, that the main-point there was not to retry the netlink
request to set the interface up. Why would that fail, and why would
a failure to set the interface up require a retry?

I think it was added to wait for carrier. But waiting for carrier was
later dropped with commit 5074898591ae99f79b0cc4872c3c7cf018abee82
and it is not clear why we would wait for carrier at all -- we don't
do that for other device types either.