wifi: clean up clearing a pending scan
All callers of request_wireless_scan() cleared the periodic scan
source immediately before calling the function, so just move that
into request_wireless_scan().  The only one that doesn't clear
it is request_wireless_scan_periodic() but that sets the source
id to 0 already, so the clear has no effect.

https://mail.gnome.org/archives/networkmanager-list/2016-October/msg00012.html