build: always enable building of tests
Let's always build with tests, regardless of --enable-tests.

If the user builds with --enable-tests=no, we automake the tests
as check_PROGRAMS, which only compiles the tests during `make check`.

Otherwise, we automake them as noinst_PROGRAMS, which builds the
tests regularly as a normal build.

This way, a user who doesn't want to build the tests regularly,
still can build and run them via `make check`.