ifcfg-rh: fix signature of link_changed() callback
Depending on how arguments are passed to the called function,
this could lead to a crash.

Maybe not on 32 bit machines where the size of the pointer is
the size of an int.

Maybe not on x86_64, where the arguments are passed in registers.

Fixes: b88c309167f4073e1503bb044b0c3a40299977cd
(cherry picked from commit 548a5440e967aca7c281fda03a09d9006f394695)