ifcfg-rh: support storing newline '\n' and other ANSI control caracters
This is especially important for the team config JSON, which is expected
to contain newlines.

ANSI C quotation is bash specific, but initscripts already use #!/bin/bash.

Unfortunately, g_strescape() doesn't escape '\'' and can thus not be
used.

Also add a test that svEscape() and svUnescape() do a round-trip.

Not only consider \r and \n as candidates for ANSI C quotation, but all
ANSI control characters.