ifcfg-rh: fix svEscape() to properly handle double quoting
'\'', '~': must not be escaped with backslash.

Also, within double quotes the backslash escape character is only
removed before special caracters like '$' or '`'. Not in general.
Yes, it means that older versions of svEscape produced invalid escape
sequences that we now treat differently. But that is not realy
avoidable, it was a bug that needs to be fixed.