build: only generate documentation when BUILD_SETTING_DOCS
Generating the docs, requires man/nm-settings.xml. That is only present
when
  - ./configure --enable-gtk-doc --with-introspecton
  - in a dist-tarball, contrary to a git-tree

Only create docs, when we also regenerate the manuals (BUILD_SETTING_DOCS).
That is, you can no longer generate docs, by relying on the pre-generated
manual pages.
If you want to generate docs, you have to regenerate the manual pages
as well.

Previously, doing the following in a git-tree failed:

    $ git clean -fdx
    $ ./autogen.sh --enable-gtk-doc --enable-introspection=no && make
    ...
    make[2]: Entering directory './NetworkManager/docs/api'
      DOC   Preparing build
      DOC   Scanning header files
      DOC   Introspecting gobjects
      DOC   Building XML
      DOC   Building XML
    make[2]: *** No rule to make target '../../man/nm-settings.xml', needed by 'html-build.stamp'.  Stop.
    make[2]: Leaving directory './NetworkManager/docs/api'