Add hint to configuration parameter in resolv.conf header
After setting the hostname with hostnamectl NetworkManager got triggered
and created a empty resolv.conf because no DNS servers had been specified in
ifcfg-* files (Fedora/EL).

As we manage it with a config management tool I wanted to disable this.
It took me some time to discover how to do it. This adds a hint to the
generated where to look for it.

Ref https://bugzilla.redhat.com/show_bug.cgi?id=1344303